Time: 3 ms (3.10%) | Memory: 8.5 MB (70.30%) - LeetSync

+++ 50-powx-n/powx-n.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     double myPow(double x, int n) {
-        long long N = n;  // To safely handle INT_MIN
+        double N = n;  // To safely handle INT_MIN
 
         if (N < 0) {
             x = 1 / x;
